requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
openai
python-dotenv
